+++ source/engine/lynxchan/comment/parseComment.js
@@ -22,11 +22,9 @@ export default function parseComment(post, {
 	defaultAuthorName,
 	capcode
 }) {
+	// Fixes LynxChan new line characters.
 	// `post.markdown` is not really "markdown", it's HTML.
-	// `lynxchan` has a bug of inserting "carriage return" (U+000D)
-	// characters before every "new line" (<br>).
-	// This workaround fixes that:
-	const content = post.markdown.replace(/\u000d/g, '')
+	const content = fixNewLineCharacters(post.markdown)
 	const authorRole = parseAuthorRole(post, { capcode })
 	const author = parseAuthor(post.name, { defaultAuthorName, boardId })
 	const comment = {
@@ -89,4 +87,59 @@ export default function parseComment(post, {
 		}
 	}
 	return comment
+}
+
+/**
+ * Fixes LynxChan "new line" delimiters.
+ * Lynxchan (at least on KohlChan) has a bug
+ * of inserting "carriage return" (U+000D)
+ * characters before every "new line" (<br>).
+ * This workaround fixes that.
+ * Also, since Lynxchan 2.3 (at least on KohlChan),
+ * all line breaks have been changed from `<br/>` to `\n`.
+ * That's a weird change, but whatever,
+ * I'll just replace all `\n`s with `<br>`s.
+ */
+// Is exported only for testing.
+export function fixNewLineCharacters(html) {
+	// `lynxchan` has a bug of inserting "carriage return" (`\r`)
+	// (U+000D) (String.charCodeAt() === 13) characters before
+	// every "new line" (`\n`) (String.charCodeAt() === 10).
+	// This workaround fixes that.
+	// For example, it's relevant on KohlChan,
+	// both before and after migrating to LynxChan 2.3.
+	html = html.replace(/\u000d/g, '')
+	// On KohlChan, since migrating to LynxChan 2.3,
+	// all line breaks have been changed from `<br/>` to `\n`.
+	// That's a weird change, but whatever,
+	// I'll just replace all `\n`s with `<br>`s.
+	// (except in `<code/>` and `<span class="aa"/>` "Ascii/JIS-Art").
+	return processExcept(replaceNewLinesWithBrs, html, [{
+		tag: 'code'
+	}, {
+		tag: 'span',
+		attributes: 'class="aa"'
+	}, {
+		// Added `<pre>` just in case.
+		tag: 'pre'
+	}])
+}
+
+function replaceNewLinesWithBrs(html) {
+	return html.replace(/\n/g, '<br>')
+}
+
+function processExcept(process, html, tags) {
+	for (const { tag, attributes } of tags) {
+		const tagStartsAt = html.indexOf('<' + tag + (attributes ? ' ' + attributes : '') + '>')
+		if (tagStartsAt >= 0) {
+			const tagEndsAt = html.indexOf('</' + tag + '>', tagStartsAt)
+			if (tagEndsAt >= 0) {
+				return processExcept(process, html.slice(0, tagStartsAt), tags) +
+					html.slice(tagStartsAt, tagEndsAt + 1) +
+					processExcept(process, html.slice(tagEndsAt + 1), tags)
+			}
+		}
+	}
+	return process(html)
 }